Fix linux build

+++ include/glow/LogMessageBuilder.hpp
@@ -33,7 +33,7 @@ template <typename T>
 LogMessageBuilder& LogMessageBuilder::operator<<(const Array<T>& array)
 {
     *this << "Array(";
-    for (int i = 0; i < array.size(); ++i)
+    for (size_t i = 0; i < array.size(); ++i)
     {
         *this << array[i];
         if (i < array.size()-1)
@@ -48,7 +48,7 @@ template <typename T>
 LogMessageBuilder& LogMessageBuilder::operator<<(const std::vector<T>& vector)
 {
     *this << "vector(";
-    for (int i = 0; i < vector.size(); ++i)
+    for (size_t i = 0; i < vector.size(); ++i)
     {
         *this << vector[i];
         if (i < vector.size()-1)
@@ -63,7 +63,7 @@ template <typename T, std::size_t Count>
 LogMessageBuilder& LogMessageBuilder::operator<<(const std::array<T, Count>& array)
 {
     *this << "array(";
-    for (int i = 0; i < Count; ++i)
+    for (size_t i = 0; i < Count; ++i)
     {
         *this << array[i];
         if (i < Count-1)
